mejorando push request del codigo fuente

+++ src/main/java/com/mundo/delivery/repository/puntoDeVentaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface puntoDeVentaRepository extends JpaRepository<PuntoDeVenta, Integer> {
+public interface puntoDeVentaRepository extends JpaRepository<PuntoDeVenta, Long> {
 }
